packaging: Introduce CI job for generating hackage documentation
This adds a CI job (hackage-doc-tarball) which generates the necessary
tarballs for uploading libraries and documentation to hackage. The
release script knows to download this folder and the upload script will
also upload the release to hackage as part of the release.

The `ghc_upload_libs` script is moved from ghc-utils into .gitlab/ghc_upload_libs

There are two modes, preparation and upload.

* The `prepare` mode takes a link to a bindist and creates a folder containing the
  source and doc tarballs ready to upload to hackage.
* The `upload` mode takes the folder created by prepare and performs the upload to
  hackage.

Fixes #21493

Related to #21512

(cherry picked from commit daac5040a9656ba0ac6633c9207cfc8c169b396c)

+++ .gitlab/ci.sh
@@ -369,13 +369,17 @@ function setup_toolchain() {
 
 function cleanup_submodules() {
   start_section "clean submodules"
-  info "Cleaning submodules..."
-  # On Windows submodules can inexplicably get into funky states where git
-  # believes that the submodule is initialized yet its associated repository
-  # is not valid. Avoid failing in this case with the following insanity.
-  git submodule sync --recursive || git submodule deinit --force --all
-  git submodule update --init --recursive
-  git submodule foreach git clean -xdf
+  if [ -d .git ]; then
+    info "Cleaning submodules..."
+    # On Windows submodules can inexplicably get into funky states where git
+    # believes that the submodule is initialized yet its associated repository
+    # is not valid. Avoid failing in this case with the following insanity.
+    git submodule sync --recursive || git submodule deinit --force --all
+    git submodule update --init --recursive
+    git submodule foreach git clean -xdf
+  else
+    info "Not cleaning submodules, not in a git repo"
+  fi;
   end_section "clean submodules"
 }
